# geoFields/src/main/resources/static/OpenAPIspec.yaml
openapi: 3.0.3
info:
  title: GeoFields HTTP API
  description: |
    JSON-эндпоинты бэкенда GeoFields (поля, NDVI, менеджер организации).

    **Авторизация:** сессия Spring Security (форма логина). Запросы с браузера отправляйте с `credentials: include` / same-origin cookie.

    **CSRF:** для `POST` используется cookie `XSRF-TOKEN` и заголовок `X-XSRF-TOKEN` (см. Spring `CookieCsrfTokenRepository`).

    Главная `/` отдаёт HTML без данных пользователя; контекст сессии — `GET /api/session/context`.

    HTML-страницы (`/`, `/login`, `/register`, `/org/manager`, `/org/admin`, `/org/agronomist`) как разметка в спецификации не описаны.
  version: 1.4.3
servers:
  - url: http://localhost:8080
    description: Локальный сервер (см. `server.port` в application.properties)
...
paths:
...
  /api/org/agronomist/summary:
    get:
      tags: [OrgAgronomist]
      summary: Сводка для UI агронома
      security:
        - sessionCookie: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgronomistSummary"
        "403":
          description: Нужна роль AGRONOMIST или ORG_ADMIN
  /api/org/agronomist/fields/{fieldId}/history:
    get:
      tags: [OrgAgronomist]
      summary: История по полю
      security:
        - sessionCookie: []
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldCropHistoryItem"
        "404":
          description: Поле не найдено в организации

    post:
      tags: [OrgAgronomist]
      summary: Добавить запись истории по полю
      security:
        - sessionCookie: []
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldCropUpsertRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionMessageResponse"
        "404":
          description: Поле не найдено в организации
  /api/org/agronomist/fields/{fieldId}/status:
    put:
      tags: [OrgAgronomist]
      summary: Изменить статус поля
      description: AGRONOMIST может перевести поле в ACTIVE или OBSOLETE. Устаревшие поля скрываются из основных списков и GeoJSON-выдачи.
      security:
        - sessionCookie: []
      parameters:
        - name: fieldId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldStatusUpdateRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionMessageResponse"
        "400":
          description: Некорректный статус
        "404":
          description: Поле не найдено в организации
  /api/org/agronomist/history/{historyId}:
    put:
      tags: [OrgAgronomist]
      summary: Обновить запись истории
      security:
        - sessionCookie: []
      parameters:
        - name: historyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FieldCropUpsertRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionMessageResponse"
        "404":
          description: Запись не найдена

    delete:
      tags: [OrgAgronomist]
      summary: Удалить запись истории
      security:
        - sessionCookie: []
      parameters:
        - name: historyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionMessageResponse"
        "400":
          description: Нельзя удалить последнюю запись истории поля
        "404":
          description: Запись не найдена
  /api/org/agronomist/fields/intake:
    post:
      tags: [OrgAgronomist]
      summary: Создать новое поле и первую запись истории
      description: Создаёт поле, добавляет первую запись истории и, если контур пересекается с другими полями организации, записывает пары в таблицу `field_intersections`.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewFieldIntakeRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionMessageResponse"
        "400":
          description: Некорректная геометрия или неизвестная культура
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Сессия после входа через POST /login
  schemas:
...
    CsrfInfo:
      type: object
      nullable: true
      properties:
        parameterName:
          type: string
          nullable: true
        headerName:
          type: string
          nullable: true
          example: "X-XSRF-TOKEN"
        token:
          type: string
          nullable: true
...
    ActionMessageResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: integer
          format: int64
    FieldStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          description: ACTIVE | OBSOLETE
          enum: [ACTIVE, OBSOLETE]
    FieldCropUpsertRequest:
      type: object
      required: [cropId, cropYear]
      properties:
        cropId:
          type: integer
          format: int64
        sowingDate:
          type: string
          format: date
          nullable: true
        harvestDate:
          type: string
          format: date
          nullable: true
        sownAreaHa:
          type: number
          nullable: true
        harvestAreaHa:
          type: number
          nullable: true
        actualYield:
          type: number
          nullable: true
        totalYield:
          type: number
          nullable: true
        plannedYield:
          type: number
          nullable: true
        forecastedYield:
          type: number
          nullable: true
        sourceData:
          type: string
          nullable: true
        sowingDetails:
          type: string
          nullable: true
        cropYear:
          type: integer
    FieldCropHistoryItem:
      type: object
      properties:
        historyId:
          type: integer
          format: int64
        cropId:
          type: integer
          format: int64
        cropName:
          type: string
        sowingDate:
          type: string
          format: date
          nullable: true
        harvestDate:
          type: string
          format: date
          nullable: true
        sownAreaHa:
          type: number
          nullable: true
        harvestAreaHa:
          type: number
          nullable: true
        actualYield:
          type: number
          nullable: true
        totalYield:
          type: number
          nullable: true
        plannedYield:
          type: number
          nullable: true
        forecastedYield:
          type: number
          nullable: true
        sourceData:
          type: string
          nullable: true
        sowingDetails:
          type: string
          nullable: true
        cropYear:
          type: integer
          nullable: true
    AgronomistSummary:
      type: object
      properties:
        organizationId:
          type: integer
          format: int64
        organizationName:
          type: string
        currentLogin:
          type: string
        currentUserFullName:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldOptionItem"
        obsoleteFields:
          type: array
          items:
            $ref: "#/components/schemas/FieldOptionItem"
        crops:
          type: array
          items:
            $ref: "#/components/schemas/CropOptionItem"
        csrf:
          $ref: "#/components/schemas/CsrfInfo"
    FieldOptionItem:
      type: object
      properties:
        fieldId:
          type: integer
          format: int64
        fieldName:
          type: string
    CropOptionItem:
      type: object
      properties:
        cropId:
          type: integer
          format: int64
        cropName:
          type: string
    NewFieldIntakeRequest:
      type: object
      required: [fieldName, geometry, history]
      properties:
        fieldName:
          type: string
        geometry:
          type: object
          required: [type, coordinates]
          properties:
            type:
              type: string
            coordinates: {}
        history:
          $ref: "#/components/schemas/FieldCropUpsertRequest"